Collapse UserNameFixture into the existing User model + factory
The extra UserNameFixture/UserNameFactory pair introduced in the previous
commit duplicated UserFactory's role. Fold the missing SA-row fields
(id, first_name, last_name, payload) onto the existing User model,
extend UserFactory with the matching faker-backed classmethods, delete
tests/models/ now that the model lives alongside its factory, and
retarget every test that imported `tests.models` at `tests.factories`.
