--- tests/models/__init__.py
@@ -1,3 +1,0 @@
-from .users import User, UserNameFixture
-
-__all__ = ["User", "UserNameFixture"]
