+++ tests/unit/test_hashing.py
@@ -1,6 +1,6 @@
 from pydantic_encryption.adapters.hashing.argon2 import Argon2Adapter
 from pydantic_encryption.types import HashedValue
-from tests.models import User
+from tests.factories import User
 
 
 class TestUnitHashing:
